Everything, except messageLogger is working now
