--- src/globalVar.h
@@ -1,6 +1,0 @@
-#pragma once 
-#include "uthash.h"
-#include <string.h>
-#include "structs.h"
-#include "stack.h"
-
