+++ src/include/parserFuncs.h
@@ -1,0 +1,23 @@
+#pragma once
+#include "structs.h"
+#include "stack.h"
+
+void messageLogger(char* msg);
+int yylex(); //TODO is this necessary?
+
+int varExists(char* id);
+int funcExists(char* funcId);
+void defineFunc(char* id, char* type, int numberOfParams);
+
+void identifierDeclaration(int length, char* type);
+void pushSomething();
+void typeReplace(char** type);
+void addVariablesToFunction(char* id);
+void endScope();
+void startScope();
+void addVar(char *id, char *type, int value, int size);
+void logVars();
+void logFuncs();
+extern void yyerror (const char *msg);
+void pushSomething();
+
