+++ src/include/resource_manager.h
