+++ src/include/symboltable.h
@@ -1,0 +1,3 @@
+#pragma once
+
+void addFunc(char* id, char* type, int numberOfParams);
