+++ src/parser.y
@@ -5,10 +5,12 @@
 %{	
 #include "diag.h"
 #include "output.h"
-#include "symboltable/funcSymboltable.h"
+#include "symboltable.h"
 #include "parserFuncs.h"
 
 // Project-specific includes
+void yyerror (const char *msg);
+
 extern STACK* programstack;
 extern STRUCTFUNC* functions;
 %}
@@ -207,3 +209,4 @@ void yyerror (const char *msg)
 {
 	FATAL_COMPILER_ERROR(INVALID_SYNTAX, 0, "(%d.%d-%d.%d): %s\n", yylloc.first_line, yylloc.first_column, yylloc.last_line, yylloc.last_column, msg);
 }
+
