--- src/symboltable/funcSymboltable.c
@@ -1,12 +1,0 @@
-#include "funcSymboltable.h"
-
-#include "../structs.h"
-#include "../stack.h"
-#include "../parserFuncs.h"
-#include <string.h>
-extern STACK* programstack;
-extern STRUCTFUNC* functions;
-
-
-
-
