--- src/symboltable/varSymboltable.c
@@ -1,2 +1,0 @@
-#include "varSymboltable.h"
-
