--- src/symboltable/varSymboltable.h
@@ -1,1 +1,0 @@
-#pragma once
